Create RenameSteps2Clean
Viewing the code in the Advanced Editor yields two new lines of code at the end
of the query (wrapped lines have been indented here for ease of identification):

Following the column’s name is a new keyword you haven’t encountered yet: each. 
    The each keyword indicates to Power Query that you want this action performed for each row in the query. 
Following this, you see the List.Product() function, 
    which indicates which columns to multiply against each other and the data type the output should take.

Note: Your line breaks may not appear in the same places—or they may
not appear at all. The key pieces to remember are (1) the CalcSales line is
being treated as one complete line of code since it was split onto two lines
in the middle of the function, and (2) the line does not end with a comma
since this line precedes the in keyword.