Create SWITCH () for 1+ IF () Statements
Replicating Power Pivot’s SWITCH() Function Power Pivot has a function called SWITCH() that allows you to
perform multi-condition logic by declaring a table of index values and results and then passing in a variable
index value. The function then looks up the provided index value and returns the matching result. 

This is easier to maintain than several levels of nested IF() statements, 
so it sometimes makes sense to replicate the SWITCH() function in Power Query.